refactor: rename rendering→render, rotation→rotate, strip raw_lines_ prefix
