+++ src/lib.rs
@@ -1,10 +1,10 @@
 mod diagram;
 mod moves;
 mod raw_lines;
-mod rendering;
-mod rotation;
+mod render;
+mod rotate;
 
 pub use diagram::{AbbreviatedDiagram, AbbreviatedItem};
 pub use diagram::{ascii_print, ascii_print_compact, try_ascii_print, try_ascii_print_compact};
 pub use moves::{DiagramMove, DiagramMoves, Lean, Move, OverUnder, UpDown};
-pub use rendering::{Horiz, VerboseDiagram, VerboseLine};
+pub use render::{Horiz, VerboseDiagram, VerboseLine};
